fix: Corrigindo tipo do campo price

+++ src/main/java/io/github/alvdelci/libraryapi/model/Book.java
@@ -3,6 +3,7 @@ package io.github.alvdelci.libraryapi.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -29,8 +30,7 @@ public class Book {
     private BookGender gender;
 
     @Column(name = "price", precision = 18, scale = 2)
-    private Double price;
-    //private BigDecimal price;
+    private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "author_id")
